+++ scripts/jp.ts
@@ -1,0 +1,111 @@
+#! /usr/bin/env node
+
+'use strict';
+
+import jmespath from '../dist/jmespath.umd';
+import { parseArgs, ParseArgsConfig } from 'node:util';
+import * as fs from 'fs';
+
+const args = getArgs();
+
+
+if (args.values.help) {
+  printHelp();
+  process.exit(0);
+}
+
+if (!args.values['expr-file'] && args.positionals.length < 1) {
+  console.log('Must provide a jmespath expression.');
+  process.exit(1);
+}
+
+
+var expression = '';
+if (args.values['expr-file']) {
+  expression = fs.readFileSync(<string>args.values['expr-file'], { encoding: 'utf8', flag: 'r' });
+} else {
+  expression = args.positionals[0];
+}
+
+
+var inputJSON = '';
+if (args?.values?.filename) {
+  inputJSON = fs.readFileSync(<string>args.values.filename, { encoding: 'utf8', flag: 'r' });
+  printResult(inputJSON, expression, <boolean>args.values.compact);
+
+} else {
+
+  process.stdin.setEncoding('utf-8');
+  process.stdin.on('readable', function () {
+    var chunk = process.stdin.read();
+    if (chunk !== null) {
+      inputJSON += chunk;
+    }
+  });
+
+  process.stdin.on('end', function () {
+    printResult(inputJSON, expression, <boolean>args.values.compact);
+  });
+}
+
+
+function getArgs() {
+  const config: ParseArgsConfig = {
+    options: {
+      compact: {
+        type: 'boolean',
+        short: 'c',
+        default: false,
+      },
+      help: {
+        type: 'boolean',
+        short: 'h',
+        default: false,
+      },
+      filename: {
+        type: 'string',
+        short: 'f',
+      },
+      'expr-file': {
+        type: 'string',
+        short: 'e',
+      }
+
+    },
+    allowPositionals: true,
+  }
+
+  return parseArgs(config);
+}
+
+
+function printHelp(): void {
+  const pkg = require('../package.json');
+  console.log(`
+    NAME:
+    jp - jp [<options>] <expression>
+
+      USAGE:
+      jp [global options] command [command options] [arguments...]
+
+      VERSION:
+      ${pkg.name}@${pkg.version}
+
+      OPTIONS:
+      --compact, -c                Produce compact JSON output that omits nonessential whitespace.
+      --filename value, -f value   Read input JSON from a file instead of stdin.
+      --expr-file value, -e value  Read JMESPath expression from the specified file.
+      --help, -h                   Show help
+      `);
+}
+
+
+function printResult(inputJSON: string, expression: string, compact: boolean = false) {
+  try {
+    var parsedInput = JSON.parse(inputJSON);
+    console.log(JSON.stringify(jmespath.search(parsedInput, expression, null), null, compact ? 0 : 2));
+  } catch (e) {
+    console.error('Error parsing JSON input:', e.message);
+    process.exit(1);
+  }
+}
